requirements.txt
PyYAML==3.10
docopt==0.4.0
requests==0.13.1
boto==2.5.1
